WIP ufbx

+++ include/pr/view3d-12/model/skeleton.h
@@ -12,21 +12,21 @@ namespace pr::rdr12
 	struct Skeleton: RefCounted<Skeleton>
 	{
 		// See description in "animation.h"
-		using Ids = pr::vector<uint64_t>;
+		using Ids = pr::vector<uint32_t>;
 		using Bones = pr::vector<m4x4>;
 		using Names = pr::vector<string32>;
 		using Hierarchy = pr::vector<uint8_t>;
 
-		uint64_t m_id;         // A unique ID for the skeleton
+		uint32_t m_id;         // A unique ID for the skeleton
 		Ids m_bone_ids;        // A unique ID for each bone
 		Names m_names;         // A name for each bone (debugging mostly)
 		Bones m_o2bp;          // The inverse of the bind-pose to object-space transform for each bone
 		Hierarchy m_hierarchy; // Depth-first ordered list of bones. First == root == 0.
 
-		Skeleton(uint64_t id, std::span<uint64_t const> bone_ids, std::span<string32 const> names, std::span<m4x4 const> o2bp, std::span<uint8_t const> hierarchy);
+		Skeleton(uint32_t id, std::span<uint32_t const> bone_ids, std::span<string32 const> names, std::span<m4x4 const> o2bp, std::span<uint8_t const> hierarchy);
 
 		// The ID of the root bone
-		uint64_t Id() const;
+		uint32_t Id() const;
 
 		// The number of bones in this skeleton
 		int BoneCount() const;
